# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Offix API",
    "description": "API REST del backend de Offix, red social de oficios (UTN FRSF).\n\n### Cómo autenticarse\n1. `POST /api/v1/auth/registro` para crear un usuario.\n2. `POST /api/v1/auth/login` para obtener un `access_token` (JWT).\n3. Enviarlo en cada request protegida como header `Authorization: Bearer <access_token>` (en Swagger UI: botón **Authorize**, arriba a la derecha).\n\n### Para el equipo de QA (Hoppscotch)\nEl spec completo también está disponible en `/openapi.json` y como archivo estático en `docs/openapi.json` del repo — se puede importar directo en Hoppscotch como colección.\n\n### Estado del proyecto\nSprint 1 — flujo de reseña (HU-01) con validación de enlace simulada, sin envío real de mail/WhatsApp. Ver `docs/plan-sprint-1.md` para el detalle de alcance y decisiones.",
    "contact": {
      "name": "Equipo Offix - UTN FRSF"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Entorno local (iniciar_backend.bat)"
    }
  ],
  "paths": {
...
    "/api/v1/oferentes/{oferente_id}/solicitudes-resena": {
      "post": {
        "tags": [
          "Reseñas",
          "Reseñas"
        ],
        "summary": "Generar enlace de reseña (paso 1 del flujo HU-01)",
        "description": "RF10 — El Oferente (logueado) genera el link/QR único de reseña, indicando\nel contacto de referencia del cliente al que se lo va a enviar.\n\nSprint 1: el envío del enlace es **simulado** — no se manda mail/WhatsApp\nreal todavía. El `codigo_unico` devuelto en la respuesta es el dato que hay\nque pegar en `POST /api/v1/resenas` (o codificar en un QR) para simular el\nlink que recibiría el cliente.",
        "operationId": "generar_solicitud_resena_api_v1_oferentes__oferente_id__solicitudes_resena_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "oferente_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Oferente Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudResenaCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolicitudResenaOut"
                }
              }
            }
          },
          "401": {
            "description": "Falta token o es inválido"
          },
          "403": {
            "description": "El oferente autenticado no es el dueño del perfil"
          },
          "404": {
            "description": "Oferente no encontrado"
          },
          "422": {
            "description": "Falta contacto_referencia_cliente o es inválido"
          }
        }
      }
    },
    "/api/v1/resenas": {
      "post": {
        "tags": [
          "Reseñas"
        ],
        "summary": "Registrar una reseña (paso 2 del flujo HU-01, público, sin login)",
        "description": "RF10 — El cliente deja la reseña accediendo exclusivamente vía el\n`codigo_unico` recibido en el enlace/QR, sin necesidad de loguearse. Queda\nen estado `Pendiente_Aprobacion` hasta que el Oferente la modere\n(`PATCH /resenas/{id}/moderar`).",
        "operationId": "registrar_resena_api_v1_resenas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResenaCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResenaOut"
                }
              }
            }
          },
          "400": {
            "description": "codigo_unico inexistente o ya utilizado"
          },
          "422": {
            "description": "Falta un campo obligatorio o un criterio está fuera de 1-5"
          }
        }
      }
    },
    "/api/v1/oferentes/{oferente_id}/resenas": {
      "get": {
        "tags": [
          "Reseñas"
        ],
        "summary": "Listar reseñas aprobadas de un oferente (público)",
        "description": "RF7 — Solo devuelve reseñas en estado `Aprobada`; las pendientes y\nrechazadas no son visibles públicamente.",
        "operationId": "listar_resenas_publicas_api_v1_oferentes__oferente_id__resenas_get",
        "parameters": [
          {
            "name": "oferente_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Oferente Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResenaOut"
                  },
                  "title": "Response Listar Resenas Publicas Api V1 Oferentes  Oferente Id  Resenas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resenas/{resena_id}/moderar": {
      "patch": {
        "tags": [
          "Reseñas"
        ],
        "summary": "Aprobar o rechazar una reseña pendiente",
        "description": "RF11/RF13 — El Oferente aprueba o rechaza la reseña. Si se rechazan 5 o\nmás de las últimas 10 reseñas del oferente, se genera automáticamente una\n`AlertaAdministrador` (RF11) visible en `GET /admin/alertas`.",
        "operationId": "moderar_resena_api_v1_resenas__resena_id__moderar_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "resena_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resena Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResenaModeracion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResenaOut"
                }
              }
            }
          },
          "400": {
            "description": "La reseña ya fue moderada (no está Pendiente_Aprobacion)"
          },
          "401": {
            "description": "Falta token o es inválido"
          },
          "403": {
            "description": "El oferente autenticado no es el dueño de la reseña"
          },
          "404": {
            "description": "Reseña no encontrada"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/alertas": {
      "get": {
        "tags": [
          "Administración"
        ],
        "summary": "Listar alertas generadas por rechazos reiterados de reseñas",
        "description": "RF17 — Revisión de alertas por rechazos reiterados de reseñas.",
        "operationId": "listar_alertas_api_v1_admin_alertas_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Falta token o es inválido"
          },
          "403": {
            "description": "Requiere rol Administrador"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "CalificacionesComentariosIn": {
        "properties": {
          "puntuacion_global": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Puntuacion Global",
            "description": "Puntuación global de la reseña, 1 a 5."
          },
          "criterios": {
            "$ref": "#/components/schemas/CriteriosValoracion"
          },
          "comentario": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comentario",
            "description": "Comentario de texto libre, opcional."
          }
        },
        "type": "object",
        "required": [
          "puntuacion_global",
          "criterios"
        ],
        "title": "CalificacionesComentariosIn",
        "example": {
          "comentario": "Excelente trabajo, muy prolijo y llegó puntual.",
          "criterios": {
            "atencion": 5,
            "calidad": 5,
            "precio": 4,
            "puntualidad": 4
          },
          "puntuacion_global": 5
        }
      },
...
      "CriteriosValoracion": {
        "properties": {
          "precio": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Precio",
            "description": "Puntuación 1-5 por el precio del servicio."
          },
          "calidad": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Calidad",
            "description": "Puntuación 1-5 por la calidad/ejecución del trabajo."
          },
          "atencion": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Atencion",
            "description": "Puntuación 1-5 por el trato/atención al cliente."
          },
          "puntualidad": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Puntualidad",
            "description": "Puntuación 1-5 por la puntualidad/cumplimiento de horarios."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "precio",
          "calidad",
          "atencion",
          "puntualidad"
        ],
        "title": "CriteriosValoracion",
        "example": {
          "atencion": 5,
          "calidad": 5,
          "precio": 4,
          "puntualidad": 4
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResenaCreate": {
        "properties": {
          "codigo_unico": {
            "type": "string",
            "title": "Codigo Unico",
            "description": "Código recibido en el enlace/QR generado por el Oferente."
          },
          "nombre_cliente": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Nombre Cliente"
          },
          "contacto_cliente_ingresado": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Contacto Cliente Ingresado",
            "description": "WhatsApp o email del cliente que deja la reseña."
          },
          "calificaciones_comentarios": {
            "$ref": "#/components/schemas/CalificacionesComentariosIn"
          }
        },
        "type": "object",
        "required": [
          "codigo_unico",
          "nombre_cliente",
          "contacto_cliente_ingresado",
          "calificaciones_comentarios"
        ],
        "title": "ResenaCreate",
        "example": {
          "calificaciones_comentarios": {
            "comentario": "Excelente trabajo, muy prolijo y llegó puntual.",
            "criterios": {
              "atencion": 5,
              "calidad": 5,
              "precio": 4,
              "puntualidad": 4
            },
            "puntuacion_global": 5
          },
          "codigo_unico": "a1b2c3d4e5f6...",
          "contacto_cliente_ingresado": "juan.cliente@example.com",
          "nombre_cliente": "Juan Cliente"
        }
      },
      "ResenaModeracion": {
        "properties": {
          "aprobar": {
            "type": "boolean",
            "title": "Aprobar",
            "description": "true = aprobar y publicar la reseña; false = rechazarla."
          }
        },
        "type": "object",
        "required": [
          "aprobar"
        ],
        "title": "ResenaModeracion",
        "example": {
          "aprobar": true
        }
      },
      "ResenaOut": {
        "properties": {
          "id_resena": {
            "type": "integer",
            "title": "Id Resena"
          },
          "oferente_id": {
            "type": "integer",
            "title": "Oferente Id"
          },
          "solicitud_id": {
            "type": "integer",
            "title": "Solicitud Id"
          },
          "nombre_cliente": {
            "type": "string",
            "title": "Nombre Cliente"
          },
          "calificaciones_comentarios": {
            "type": "object",
            "title": "Calificaciones Comentarios"
          },
          "estado": {
            "type": "string",
            "title": "Estado",
            "description": "Pendiente_Aprobacion | Aprobada | Rechazada"
          },
          "fecha_creacion": {
            "type": "string",
            "format": "date-time",
            "title": "Fecha Creacion"
          }
        },
        "type": "object",
        "required": [
          "id_resena",
          "oferente_id",
          "solicitud_id",
          "nombre_cliente",
          "calificaciones_comentarios",
          "estado",
          "fecha_creacion"
        ],
        "title": "ResenaOut"
      },
...
      "SolicitudResenaCreate": {
        "properties": {
          "contacto_referencia_cliente": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Contacto Referencia Cliente",
            "description": "Contacto (nombre y WhatsApp o email) del cliente al que se le va a enviar el enlace de reseña."
          }
        },
        "type": "object",
        "required": [
          "contacto_referencia_cliente"
        ],
        "title": "SolicitudResenaCreate",
        "example": {
          "contacto_referencia_cliente": "Juan Cliente - +54 9 11 5555-1234"
        }
      },
      "SolicitudResenaOut": {
        "properties": {
          "id_solicitud": {
            "type": "integer",
            "title": "Id Solicitud"
          },
          "oferente_id": {
            "type": "integer",
            "title": "Oferente Id"
          },
          "codigo_unico": {
            "type": "string",
            "title": "Codigo Unico",
            "description": "Código único a incluir en el link/QR que se comparte con el cliente."
          },
          "contacto_referencia_cliente": {
            "type": "string",
            "title": "Contacto Referencia Cliente"
          },
          "estado": {
            "type": "string",
            "title": "Estado",
            "description": "Pendiente_Uso | Utilizada | Expirada"
          },
          "fecha_generacion": {
            "type": "string",
            "format": "date-time",
            "title": "Fecha Generacion"
          },
          "fecha_expiracion": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Expiracion"
          }
        },
        "type": "object",
        "required": [
          "id_solicitud",
          "oferente_id",
          "codigo_unico",
          "contacto_referencia_cliente",
          "estado",
          "fecha_generacion"
        ],
        "title": "SolicitudResenaOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}